git-svn-id: https://svn.cs.biu.ac.il/svn/development/SDK@1151 ba1fec17-5bc3-4c1d-a110-6a66ae20c876

+++ Code/branches/NoInitParams/biu/scapi/primitives/dlog/DlogGroup.java
@@ -146,7 +146,7 @@ public interface DlogGroup {
 	 * @param exponent
 	 * @return the exponentiation result
 	 */
-	public GroupElement exponentiateWithPreComputedValues(GroupElement groupElement, int exponent);
+	public GroupElement exponentiateWithPreComputedValues(GroupElement groupElement, BigInteger exponent);
 	
 	/**
 	 * Converts a byte array to a GroupElement.
